Move root config to signal request.
