+++ server/neptune/workflows/deploy.go
@@ -6,24 +6,25 @@ import (
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/deploy"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/deploy/revision"
+	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/deploy/request"
 	"github.com/uber-go/tally/v4"
 	"go.temporal.io/sdk/workflow"
 )
 
 // Export anything that callers need such as requests, signals, etc.
 type DeployRequest = deploy.Request
-type Repo = deploy.Repo
-type Root = deploy.Root
-type Job = deploy.Job
-type Step = deploy.Step
-type AppCredentials = deploy.AppCredentials
-type HeadCommit = deploy.Commit
-type Ref = deploy.Ref
+type Repo = request.Repo
+type Root = request.Root
+type Job = request.Job
+type Step = request.Step
+type AppCredentials = request.AppCredentials
+type HeadCommit = request.Commit
+type Ref = request.Ref
 
-type PlanMode = deploy.PlanMode
+type PlanMode = request.PlanMode
 
-const DestroyPlanMode = deploy.DestroyPlanMode
-const NormalPlanMode = deploy.NormalPlanMode
+const DestroyPlanMode = request.DestroyPlanMode
+const NormalPlanMode = request.NormalPlanMode
 
 type DeployNewRevisionSignalRequest = revision.NewRevisionRequest
